refactor: extract tilt relaxation from Minimizer to TiltRelaxationManager (Phase 2)
