+++ runtime/minimizer.py
@@ -8,11 +8,7 @@ import numpy as np
 
 from core.parameters.global_parameters import GlobalParameters
 from core.parameters.resolver import ParameterResolver
-from geometry.entities import Mesh, _fast_cross
-from modules.energy.leaflet_presence import (
-    leaflet_absent_vertex_mask,
-    leaflet_present_triangle_mask,
-)
+from geometry.entities import Mesh
 from runtime.constraint_manager import ConstraintModuleManager
 from runtime.diagnostics.audit import (
     check_gauss_bonnet,
@@ -34,88 +30,14 @@ from runtime.minimizer_helpers import (
     build_reduced_line_search_trial_energy_fn,
     get_cached_tilt_fixed_mask,
 )
-from runtime.preconditioners import (
-    build_leaflet_tilt_cg_preconditioner,
-    build_tilt_cg_preconditioner,
-)
 from runtime.projections.curved_disk import project_curved_free_disk_shape_dofs
 from runtime.projections.tilt import (
-    build_leaflet_trial_tilts,
-    project_leaflet_tilts_with_optional_axisymmetry,
     project_tilts_to_tangent_array,
 )
 from runtime.steppers.base import BaseStepper
+from runtime.steppers.tilt_relaxation import TiltRelaxationManager
 
 logger = logging.getLogger("membrane_solver")
-
-
-def _apply_inner_coupled_update_mode_to_delta(
-    *,
-    mesh: Mesh,
-    global_params: GlobalParameters,
-    positions: np.ndarray,
-    fixed_mask_in: np.ndarray,
-    delta_in: np.ndarray,
-) -> tuple[np.ndarray, dict[str, float | int | str | bool]]:
-    """Modify generated inner trial updates in benchmark-only continuation modes."""
-    mode = str(global_params.get("inner_coupled_update_mode") or "off").strip().lower()
-    if mode not in {"off", "rim_matched_radial_continuation_v1"}:
-        raise ValueError(
-            "inner_coupled_update_mode must be 'off' or "
-            "'rim_matched_radial_continuation_v1'."
-        )
-    radius = float(global_params.get("benchmark_disk_radius") or 0.0)
-    lambda_value = float(global_params.get("benchmark_lambda_value") or 0.0)
-    stats: dict[str, float | int | str | bool] = {
-        "enabled": bool(mode != "off"),
-        "mode": str(mode),
-        "candidate_row_count": 0,
-        "capped_row_count": 0,
-        "rim_row_count": 0,
-        "cap_magnitude": 0.0,
-    }
-    if mode == "off" or radius <= 0.0 or lambda_value <= 0.0:
-        return delta_in, stats
-
-    from modules.constraints.local_interface_shells import radial_unit_vectors
-    from modules.energy.bending_tilt_leaflet import _assume_J0_center_xy
-
-    center_xy = _assume_J0_center_xy(global_params)
-    shifted = np.array(positions, copy=True)
-    shifted[:, 0] = shifted[:, 0] - float(center_xy[0])
-    shifted[:, 1] = shifted[:, 1] - float(center_xy[1])
-    radii, r_hat = radial_unit_vectors(shifted)
-    rim_w = float(lambda_value)
-    near_w = float(4.0 * lambda_value)
-    free_mask = ~np.asarray(fixed_mask_in, dtype=bool)
-    rim_rows = np.flatnonzero((np.abs(radii - radius) <= rim_w) & free_mask)
-    target_rows = np.flatnonzero(
-        (radii > (radius + rim_w)) & (radii <= (radius + near_w)) & free_mask
-    )
-    stats["candidate_row_count"] = int(target_rows.size)
-    stats["rim_row_count"] = int(rim_rows.size)
-    if rim_rows.size == 0 or target_rows.size == 0:
-        return delta_in, stats
-
-    rim_delta_rad = np.einsum("ij,ij->i", delta_in[rim_rows], r_hat[rim_rows])
-    cap_mag = (
-        float(1.05 * np.median(np.abs(rim_delta_rad))) if rim_delta_rad.size else 0.0
-    )
-    stats["cap_magnitude"] = float(cap_mag)
-    if cap_mag <= 0.0:
-        return delta_in, stats
-
-    out = np.array(delta_in, copy=True)
-    target_delta_rad = np.einsum("ij,ij->i", out[target_rows], r_hat[target_rows])
-    capped = np.clip(target_delta_rad, -cap_mag, cap_mag)
-    adjust = capped - target_delta_rad
-    hit = np.abs(adjust) > 1.0e-14
-    if not np.any(hit):
-        return out, stats
-    rows_hit = target_rows[hit]
-    out[rows_hit] += adjust[hit][:, None] * r_hat[rows_hit]
-    stats["capped_row_count"] = int(np.sum(hit))
-    return out, stats
 
 
 class Minimizer:
@@ -210,6 +132,24 @@ class Minimizer:
             "rim_row_count": 0,
             "cap_magnitude": 0.0,
         }
+
+        self._tilt_relaxation_manager = TiltRelaxationManager(
+            mesh=self.mesh,
+            global_params=self.global_params,
+            param_resolver=self.param_resolver,
+            constraint_manager=self.constraint_manager,
+            energy_context_fn=self.energy_context,
+            compute_energy_and_tilt_gradient_array_fn=self._compute_energy_and_tilt_gradient_array,
+            compute_energy_array_with_tilts_fn=self._compute_energy_array_with_tilts,
+            compute_energy_and_leaflet_tilt_gradients_array_fn=self._compute_energy_and_leaflet_tilt_gradients_array,
+            compute_tilt_dependent_energy_with_leaflet_tilts_fn=self._compute_tilt_dependent_energy_with_leaflet_tilts,
+            set_leaflet_tilts_from_arrays_fast_fn=self._set_leaflet_tilts_from_arrays_fast,
+            triangle_rows_fn=self._triangle_rows,
+            tilt_fixed_mask_fn=self._tilt_fixed_mask,
+            tilt_fixed_mask_in_fn=self._tilt_fixed_mask_in,
+            tilt_fixed_mask_out_fn=self._tilt_fixed_mask_out,
+            constraint_modules=self.constraint_modules,
+        )
 
     def reset_soa_caches(self) -> None:
         """Clear cached SOA views after mesh-replacing operations."""
@@ -1127,23 +1067,6 @@ class Minimizer:
 
         return float(total_energy)
 
-    @staticmethod
-    def _tilt_vertex_areas_from_triangles(
-        *, n_vertices: int, tri_rows: np.ndarray, positions: np.ndarray
-    ) -> np.ndarray:
-        """Return barycentric per-vertex areas based on triangle areas."""
-        tri_pos = positions[tri_rows]
-        v0 = tri_pos[:, 0, :]
-        v1 = tri_pos[:, 1, :]
-        v2 = tri_pos[:, 2, :]
-        areas = 0.5 * np.linalg.norm(_fast_cross(v1 - v0, v2 - v0), axis=1)
-        vertex_areas = np.zeros(n_vertices, dtype=float)
-        a3 = areas / 3.0
-        np.add.at(vertex_areas, tri_rows[:, 0], a3)
-        np.add.at(vertex_areas, tri_rows[:, 1], a3)
-        np.add.at(vertex_areas, tri_rows[:, 2], a3)
-        return vertex_areas
-
     def _compute_energy_and_leaflet_tilt_gradients_array(
         self,
         *,
@@ -1262,187 +1185,7 @@ class Minimizer:
         mode: str,
     ) -> None:
         """Relax vertex tilt vectors according to the configured solve mode."""
-        mode_norm = str(mode or "").strip().lower()
-        if mode_norm in ("", "none", "off", "false", "fixed"):
-            return
-
-        if mode_norm not in ("nested", "coupled"):
-            logger.warning("Unknown tilt_solve_mode=%r; treating as 'fixed'.", mode)
-            return
-
-        step_size = float(self.global_params.get("tilt_step_size", 0.0) or 0.0)
-        if step_size <= 0.0:
-            return
-
-        tol = float(self.global_params.get("tilt_tol", 0.0) or 0.0)
-        if tol <= 0.0:
-            tol = 0.0
-
-        if mode_norm == "nested":
-            n_inner = int(self.global_params.get("tilt_inner_steps", 0) or 0)
-        else:
-            n_inner = int(
-                self.global_params.get(
-                    "tilt_coupled_steps", self.global_params.get("tilt_inner_steps", 0)
-                )
-                or 0
-            )
-        if n_inner <= 0:
-            return
-
-        solver = (
-            str(self.global_params.get("tilt_solver", "gd") or "gd").strip().lower()
-        )
-        if solver not in ("gd", "cg"):
-            logger.warning("Unknown tilt_solver=%r; using gradient descent.", solver)
-            solver = "gd"
-        if solver == "cg":
-            max_iters = int(self.global_params.get("tilt_cg_max_iters", n_inner) or 0)
-            if max_iters <= 0:
-                return
-        else:
-            max_iters = n_inner
-
-        fixed_mask = self._tilt_fixed_mask()
-        has_free = bool(np.any(~fixed_mask))
-        if not has_free:
-            return
-
-        with self.mesh.geometry_freeze(positions):
-            tilts = self.mesh.tilts_view().copy(order="F")
-            normals = self.mesh.vertex_normals(positions)
-            tilts = project_tilts_to_tangent_array(tilts, normals)
-            tilt_fixed_vals = tilts[fixed_mask].copy() if np.any(fixed_mask) else None
-
-            tilt_grad = np.zeros_like(tilts)
-            preconditioner = None
-            if solver == "cg":
-                preconditioner = (
-                    str(
-                        self.global_params.get("tilt_cg_preconditioner", "jacobi")
-                        or "jacobi"
-                    )
-                    .strip()
-                    .lower()
-                )
-                if preconditioner in ("none", "off", "false"):
-                    preconditioner = None
-
-            if solver == "gd":
-                for _ in range(max_iters):
-                    E0 = self._compute_energy_and_tilt_gradient_array(
-                        positions=positions, tilts=tilts, tilt_grad_arr=tilt_grad
-                    )
-                    if np.any(fixed_mask):
-                        tilt_grad[fixed_mask] = 0.0
-
-                    gnorm = float(np.linalg.norm(tilt_grad[~fixed_mask]))
-                    if gnorm == 0.0:
-                        break
-                    if tol > 0.0 and gnorm < tol:
-                        break
-
-                    step = step_size
-                    accepted = False
-                    for _bt in range(12):
-                        trial = tilts - step * tilt_grad
-                        trial = project_tilts_to_tangent_array(trial, normals)
-                        if tilt_fixed_vals is not None:
-                            trial[fixed_mask] = tilt_fixed_vals
-                        E1 = self._compute_energy_array_with_tilts(
-                            positions=positions, tilts=trial
-                        )
-                        if E1 <= E0:
-                            tilts = trial
-                            accepted = True
-                            break
-                        step *= 0.5
-                        if step < 1e-16:
-                            break
-
-                    if not accepted:
-                        break
-            else:
-                M_inv = None
-                if preconditioner == "jacobi":
-                    M_inv = build_tilt_cg_preconditioner(
-                        self.mesh,
-                        self.param_resolver,
-                        self.energy_context(),
-                        positions=positions,
-                        index_map=self.mesh.vertex_index_to_row,
-                        fixed_mask=fixed_mask,
-                    )
-
-                E0 = self._compute_energy_and_tilt_gradient_array(
-                    positions=positions, tilts=tilts, tilt_grad_arr=tilt_grad
-                )
-                if np.any(fixed_mask):
-                    tilt_grad[fixed_mask] = 0.0
-                gnorm = float(np.linalg.norm(tilt_grad[~fixed_mask]))
-                if gnorm == 0.0 or (tol > 0.0 and gnorm < tol):
-                    self.mesh.set_tilts_from_array(tilts)
-                    return
-
-                residual = -tilt_grad
-                if M_inv is not None:
-                    z_vec = residual * M_inv[:, None]
-                else:
-                    z_vec = residual
-                direction = z_vec.copy()
-                rz_old = float(np.sum(residual * z_vec))
-
-                for _ in range(max_iters):
-                    if gnorm == 0.0:
-                        break
-                    if tol > 0.0 and gnorm < tol:
-                        break
-
-                    step = step_size
-                    accepted = False
-                    for _bt in range(12):
-                        trial = tilts + step * direction
-                        trial = project_tilts_to_tangent_array(trial, normals)
-                        if tilt_fixed_vals is not None:
-                            trial[fixed_mask] = tilt_fixed_vals
-                        E1 = self._compute_energy_array_with_tilts(
-                            positions=positions, tilts=trial
-                        )
-                        if E1 <= E0:
-                            tilts = trial
-                            E0 = E1
-                            accepted = True
-                            break
-                        step *= 0.5
-                        if step < 1e-16:
-                            break
-
-                    if not accepted:
-                        break
-
-                    E0 = self._compute_energy_and_tilt_gradient_array(
-                        positions=positions, tilts=tilts, tilt_grad_arr=tilt_grad
-                    )
-                    if np.any(fixed_mask):
-                        tilt_grad[fixed_mask] = 0.0
-
-                    gnorm = float(np.linalg.norm(tilt_grad[~fixed_mask]))
-                    if gnorm == 0.0 or (tol > 0.0 and gnorm < tol):
-                        break
-
-                    residual = -tilt_grad
-                    if M_inv is not None:
-                        z_vec = residual * M_inv[:, None]
-                    else:
-                        z_vec = residual
-                    rz_new = float(np.sum(residual * z_vec))
-                    if rz_old == 0.0:
-                        break
-                    beta = rz_new / rz_old
-                    direction = z_vec + beta * direction
-                    rz_old = rz_new
-
-        self.mesh.set_tilts_from_array(tilts)
+        self._tilt_relaxation_manager.relax_tilts(positions=positions, mode=mode)
 
     def _relax_leaflet_tilts(
         self,
@@ -1451,549 +1194,15 @@ class Minimizer:
         mode: str,
     ) -> None:
         """Relax inner/outer leaflet tilt vectors according to solve mode."""
-        projection_cadence = (
-            str(
-                self.global_params.get("tilt_projection_cadence", "per_step")
-                or "per_step"
-            )
-            .strip()
-            .lower()
+        self._tilt_relaxation_manager.relax_leaflet_tilts(
+            positions=positions, mode=mode
         )
-        if projection_cadence not in {"per_step", "per_pass"}:
-            raise ValueError(
-                "tilt_projection_cadence must be 'per_step' or 'per_pass'."
-            )
-        projection_interval = int(
-            self.global_params.get("tilt_projection_interval", 1) or 1
+        self._last_tilt_projection_stats = (
+            self._tilt_relaxation_manager.last_tilt_projection_stats
         )
-        if projection_interval < 1:
-            raise ValueError("tilt_projection_interval must be >= 1.")
-        self._last_tilt_projection_stats = {
-            "projection_cadence": str(projection_cadence),
-            "projection_interval": int(projection_interval),
-            "projection_apply_count": 0,
-            "tilt_projection_norm_loss_outer_far": 0.0,
-        }
-
-        mode_norm = str(mode or "").strip().lower()
-        if mode_norm in ("", "none", "off", "false", "fixed"):
-            return
-
-        if mode_norm not in ("nested", "coupled"):
-            logger.warning("Unknown tilt_solve_mode=%r; treating as 'fixed'.", mode)
-            return
-
-        step_size = float(self.global_params.get("tilt_step_size", 0.0) or 0.0)
-        if step_size <= 0.0:
-            return
-
-        tol = float(self.global_params.get("tilt_tol", 0.0) or 0.0)
-        if tol <= 0.0:
-            tol = 0.0
-
-        if mode_norm == "nested":
-            n_inner = int(self.global_params.get("tilt_inner_steps", 0) or 0)
-        else:
-            n_inner = int(
-                self.global_params.get(
-                    "tilt_coupled_steps", self.global_params.get("tilt_inner_steps", 0)
-                )
-                or 0
-            )
-        if n_inner <= 0:
-            return
-
-        solver = (
-            str(self.global_params.get("tilt_solver", "gd") or "gd").strip().lower()
+        self._last_inner_coupled_update_mode_stats = (
+            self._tilt_relaxation_manager.last_inner_coupled_update_mode_stats
         )
-        if solver not in ("gd", "cg"):
-            logger.warning("Unknown tilt_solver=%r; using gradient descent.", solver)
-            solver = "gd"
-        if solver == "cg":
-            max_iters = int(self.global_params.get("tilt_cg_max_iters", n_inner) or 0)
-            if max_iters <= 0:
-                return
-        else:
-            max_iters = n_inner
-
-        fixed_mask_in = self._tilt_fixed_mask_in()
-        fixed_mask_out = self._tilt_fixed_mask_out()
-        has_free = bool(np.any(~fixed_mask_in) or np.any(~fixed_mask_out))
-        if not has_free:
-            return
-
-        def _apply_scaffold_interface_shape_projection() -> None:
-            mode_match = (
-                str(self.global_params.get("rim_slope_match_mode") or "")
-                .strip()
-                .lower()
-            )
-            trace_radius = self.global_params.get("parity_trace_layer_radius")
-            outer_shells = int(self.global_params.get("parity_outer_shells", 0) or 0)
-            if (
-                mode_match != "physical_edge_staggered_v1"
-                or trace_radius is None
-                or outer_shells <= 0
-            ):
-                return
-            residual_threshold = 0.1
-            for mod in self.constraint_modules:
-                if (
-                    getattr(mod, "__name__", "")
-                    != "modules.constraints.rim_slope_match_out"
-                ):
-                    continue
-                diagnostics_fn = getattr(mod, "matching_residual_diagnostics", None)
-                if not callable(diagnostics_fn):
-                    break
-                diagnostics = diagnostics_fn(
-                    self.mesh,
-                    self.global_params,
-                    self.mesh.positions_view(),
-                )
-                if not bool(diagnostics.get("available", False)):
-                    break
-                outer = diagnostics.get("outer_residual", {})
-                inner = diagnostics.get("inner_residual", {})
-                outer_max_abs = float(outer.get("max_abs", 0.0) or 0.0)
-                inner_max_abs = float(inner.get("max_abs", 0.0) or 0.0)
-                if max(outer_max_abs, inner_max_abs) <= residual_threshold:
-                    return
-                break
-            self.constraint_manager.enforce_all(
-                self.mesh,
-                global_params=self.global_params,
-                context="tilt_block",
-            )
-            self.mesh.project_tilts_to_tangent()
-            self.mesh.increment_version()
-
-        with self.mesh.geometry_freeze(positions):
-            # Ensure tilt-only constraints (e.g. thetaB boundary projection) are
-            # applied before the first tilt-gradient evaluation. Otherwise a
-            # change in a scalar tilt parameter (like thetaB) can leave the tilt
-            # field in a stale state (often all zeros), making the relaxation
-            # no-op even though constraints require a non-zero boundary tilt.
-            _apply_scaffold_interface_shape_projection()
-            if hasattr(self.constraint_manager, "enforce_tilt_constraints"):
-                self.constraint_manager.enforce_tilt_constraints(
-                    self.mesh, global_params=self.global_params
-                )
-
-            ctx = self.energy_context()
-            mesh_tilts_in = self.mesh.tilts_in_view()
-            mesh_tilts_out = self.mesh.tilts_out_view()
-            tilts_in = ctx.scratch_array(
-                "leaflet_current_tilts_in",
-                shape=mesh_tilts_in.shape,
-                dtype=mesh_tilts_in.dtype,
-            )
-            tilts_out = ctx.scratch_array(
-                "leaflet_current_tilts_out",
-                shape=mesh_tilts_out.shape,
-                dtype=mesh_tilts_out.dtype,
-            )
-            normals = self.mesh.vertex_normals(positions)
-
-            def _project_leaflet_tilts_in_place(
-                tilts_in_arr: np.ndarray, tilts_out_arr: np.ndarray
-            ) -> None:
-                dot_in = np.einsum("ij,ij->i", tilts_in_arr, normals)
-                tilts_in_arr -= dot_in[:, None] * normals
-                dot_out = np.einsum("ij,ij->i", tilts_out_arr, normals)
-                tilts_out_arr -= dot_out[:, None] * normals
-                proj_in, proj_out = project_leaflet_tilts_with_optional_axisymmetry(
-                    global_params=self.global_params,
-                    positions=positions,
-                    normals=normals,
-                    tilts_in=tilts_in_arr,
-                    tilts_out=tilts_out_arr,
-                    fixed_mask_in=fixed_mask_in,
-                    fixed_mask_out=fixed_mask_out,
-                )
-                if proj_in is not tilts_in_arr:
-                    tilts_in_arr[:] = proj_in
-                if proj_out is not tilts_out_arr:
-                    tilts_out_arr[:] = proj_out
-
-            def _load_leaflet_tilts_from_mesh(
-                tilts_in_arr: np.ndarray, tilts_out_arr: np.ndarray
-            ) -> None:
-                np.copyto(tilts_in_arr, self.mesh.tilts_in_view())
-                np.copyto(tilts_out_arr, self.mesh.tilts_out_view())
-                _project_leaflet_tilts_in_place(tilts_in_arr, tilts_out_arr)
-
-            _load_leaflet_tilts_from_mesh(tilts_in, tilts_out)
-            grad_dummy = ctx.scratch_array(
-                "leaflet_relax_grad_dummy",
-                shape=positions.shape,
-                dtype=positions.dtype,
-            )
-
-            tri_rows, _ = self._triangle_rows()
-            if tri_rows is None or len(tri_rows) == 0:
-                return
-
-            # Cache barycentric per-vertex areas once; positions are frozen for
-            # the inner loop, so pure-tilt energies/gradients can reuse these.
-            tilt_vertex_areas_in = self._tilt_vertex_areas_from_triangles(
-                n_vertices=len(self.mesh.vertex_ids),
-                tri_rows=tri_rows,
-                positions=positions,
-            )
-            absent_mask_out = leaflet_absent_vertex_mask(
-                self.mesh, self.global_params, leaflet="out"
-            )
-            tri_keep_out = leaflet_present_triangle_mask(
-                self.mesh, tri_rows, absent_vertex_mask=absent_mask_out
-            )
-            tri_rows_out = tri_rows[tri_keep_out] if tri_keep_out.size else tri_rows
-            tilt_vertex_areas_out = (
-                np.zeros(len(self.mesh.vertex_ids), dtype=float)
-                if tri_rows_out.size == 0
-                else self._tilt_vertex_areas_from_triangles(
-                    n_vertices=len(self.mesh.vertex_ids),
-                    tri_rows=tri_rows_out,
-                    positions=positions,
-                )
-            )
-            tilt_fixed_vals_in = (
-                tilts_in[fixed_mask_in].copy() if np.any(fixed_mask_in) else None
-            )
-            tilt_fixed_vals_out = (
-                tilts_out[fixed_mask_out].copy() if np.any(fixed_mask_out) else None
-            )
-
-            tilt_in_grad = ctx.scratch_array(
-                "leaflet_tilt_in_grad",
-                shape=tilts_in.shape,
-                dtype=tilts_in.dtype,
-            )
-            tilt_out_grad = ctx.scratch_array(
-                "leaflet_tilt_out_grad",
-                shape=tilts_out.shape,
-                dtype=tilts_out.dtype,
-            )
-            trial_in_buf = ctx.scratch_array(
-                "leaflet_trial_in_buf",
-                shape=tilts_in.shape,
-                dtype=tilts_in.dtype,
-            )
-            trial_out_buf = ctx.scratch_array(
-                "leaflet_trial_out_buf",
-                shape=tilts_out.shape,
-                dtype=tilts_out.dtype,
-            )
-            accepted_steps = 0
-            projection_apply_count = 0
-            projection_norm_loss_outer_far = 0.0
-            projection_norm_ref_outer_far = 0.0
-
-            outer_far_rows = np.asarray([], dtype=int)
-            projection_loss_radius = float(
-                self.global_params.get("tilt_projection_loss_radius", 0.0) or 0.0
-            )
-            projection_loss_lambda = float(
-                self.global_params.get("tilt_projection_loss_lambda", 0.0) or 0.0
-            )
-            projection_loss_outer_near_width = float(
-                self.global_params.get(
-                    "tilt_projection_loss_outer_near_width_lambda", 4.0
-                )
-                or 4.0
-            )
-            if projection_loss_radius > 0.0 and projection_loss_lambda > 0.0:
-                rxy = np.linalg.norm(positions[:, :2], axis=1)
-                outer_far_rows = np.flatnonzero(
-                    rxy
-                    >= (
-                        projection_loss_radius
-                        + projection_loss_outer_near_width * projection_loss_lambda
-                    )
-                )
-
-            def _refresh_tilts_from_constraints_and_project() -> None:
-                nonlocal tilts_in, tilts_out
-                nonlocal projection_apply_count
-                nonlocal projection_norm_loss_outer_far, projection_norm_ref_outer_far
-
-                before_norm = None
-                if outer_far_rows.size > 0:
-                    before_norm = np.linalg.norm(tilts_in[outer_far_rows], axis=1)
-                self._set_leaflet_tilts_from_arrays_fast(tilts_in, tilts_out)
-                if hasattr(self.constraint_manager, "enforce_tilt_constraints"):
-                    self.constraint_manager.enforce_tilt_constraints(
-                        self.mesh, global_params=self.global_params
-                    )
-                _load_leaflet_tilts_from_mesh(tilts_in, tilts_out)
-                projection_apply_count += 1
-                if before_norm is not None:
-                    after_norm = np.linalg.norm(tilts_in[outer_far_rows], axis=1)
-                    projection_norm_loss_outer_far += float(
-                        np.sum(np.maximum(before_norm - after_norm, 0.0))
-                    )
-                    projection_norm_ref_outer_far += float(np.sum(before_norm))
-
-            def _leaflet_tilt_gradients() -> tuple[float, float]:
-                E0 = self._compute_energy_and_leaflet_tilt_gradients_array(
-                    positions=positions,
-                    tilts_in=tilts_in,
-                    tilts_out=tilts_out,
-                    tilt_in_grad_arr=tilt_in_grad,
-                    tilt_out_grad_arr=tilt_out_grad,
-                    tilt_vertex_areas_in=tilt_vertex_areas_in,
-                    tilt_vertex_areas_out=tilt_vertex_areas_out,
-                    grad_dummy=grad_dummy,
-                    tilt_only=True,
-                )
-                if hasattr(
-                    self.constraint_manager, "apply_tilt_gradient_modifications_array"
-                ):
-                    self.constraint_manager.apply_tilt_gradient_modifications_array(
-                        tilt_in_grad,
-                        tilt_out_grad,
-                        self.mesh,
-                        self.global_params,
-                        positions=positions,
-                        tilts_in=tilts_in,
-                        tilts_out=tilts_out,
-                    )
-                if np.any(fixed_mask_in):
-                    tilt_in_grad[fixed_mask_in] = 0.0
-                if np.any(fixed_mask_out):
-                    tilt_out_grad[fixed_mask_out] = 0.0
-
-                gnorm = float(
-                    np.sqrt(
-                        np.sum(tilt_in_grad[~fixed_mask_in] ** 2)
-                        + np.sum(tilt_out_grad[~fixed_mask_out] ** 2)
-                    )
-                )
-                return float(E0), gnorm
-
-            preconditioner = None
-            if solver == "cg":
-                preconditioner = (
-                    str(
-                        self.global_params.get("tilt_cg_preconditioner", "jacobi")
-                        or "jacobi"
-                    )
-                    .strip()
-                    .lower()
-                )
-                if preconditioner in ("none", "off", "false"):
-                    preconditioner = None
-
-            if solver == "gd":
-                for _ in range(max_iters):
-                    E0, gnorm = _leaflet_tilt_gradients()
-                    if gnorm == 0.0:
-                        break
-                    if tol > 0.0 and gnorm < tol:
-                        break
-
-                    step = step_size
-                    accepted = False
-                    for _bt in range(12):
-                        delta_in_trial = -step * tilt_in_grad
-                        delta_in_trial, update_mode_stats = (
-                            _apply_inner_coupled_update_mode_to_delta(
-                                mesh=self.mesh,
-                                global_params=self.global_params,
-                                positions=positions,
-                                fixed_mask_in=fixed_mask_in,
-                                delta_in=delta_in_trial,
-                            )
-                        )
-                        self._last_inner_coupled_update_mode_stats = dict(
-                            update_mode_stats
-                        )
-                        trial_in, trial_out = build_leaflet_trial_tilts(
-                            base_in=tilts_in,
-                            base_out=tilts_out,
-                            delta_in=delta_in_trial,
-                            delta_out=-step * tilt_out_grad,
-                            normals=normals,
-                            fixed_mask_in=fixed_mask_in,
-                            fixed_mask_out=fixed_mask_out,
-                            fixed_vals_in=tilt_fixed_vals_in,
-                            fixed_vals_out=tilt_fixed_vals_out,
-                            out_in=trial_in_buf,
-                            out_out=trial_out_buf,
-                        )
-                        E1 = self._compute_tilt_dependent_energy_with_leaflet_tilts(
-                            positions=positions,
-                            tilts_in=trial_in,
-                            tilts_out=trial_out,
-                            grad_dummy=grad_dummy,
-                            tilt_vertex_areas_in=tilt_vertex_areas_in,
-                            tilt_vertex_areas_out=tilt_vertex_areas_out,
-                        )
-                        if E1 <= E0:
-                            tilts_in, trial_in_buf = trial_in, tilts_in
-                            tilts_out, trial_out_buf = trial_out, tilts_out
-                            accepted = True
-                            break
-                        step *= 0.5
-                        if step < 1e-16:
-                            break
-
-                    if not accepted:
-                        break
-
-                    accepted_steps += 1
-                    if (
-                        projection_cadence == "per_step"
-                        and (accepted_steps % projection_interval) == 0
-                    ):
-                        _refresh_tilts_from_constraints_and_project()
-            else:
-                M_inv_in = None
-                M_inv_out = None
-                if preconditioner == "jacobi":
-                    (
-                        M_inv_in,
-                        M_inv_out,
-                    ) = build_leaflet_tilt_cg_preconditioner(
-                        self.mesh,
-                        self.param_resolver,
-                        self.energy_context(),
-                        positions=positions,
-                        index_map=self.mesh.vertex_index_to_row,
-                        fixed_mask_in=fixed_mask_in,
-                        fixed_mask_out=fixed_mask_out,
-                    )
-
-                E0, gnorm = _leaflet_tilt_gradients()
-                if gnorm == 0.0 or (tol > 0.0 and gnorm < tol):
-                    self._set_leaflet_tilts_from_arrays_fast(tilts_in, tilts_out)
-                    return
-
-                res_in = -tilt_in_grad
-                res_out = -tilt_out_grad
-                if M_inv_in is not None:
-                    z_in = res_in * M_inv_in[:, None]
-                else:
-                    z_in = res_in
-                if M_inv_out is not None:
-                    z_out = res_out * M_inv_out[:, None]
-                else:
-                    z_out = res_out
-
-                dir_in = z_in.copy()
-                dir_out = z_out.copy()
-                rz_old = float(np.sum(res_in * z_in) + np.sum(res_out * z_out))
-
-                for _ in range(max_iters):
-                    if gnorm == 0.0:
-                        break
-                    if tol > 0.0 and gnorm < tol:
-                        break
-
-                    step = step_size
-                    accepted = False
-                    for _bt in range(12):
-                        delta_in_trial = step * dir_in
-                        delta_in_trial, update_mode_stats = (
-                            _apply_inner_coupled_update_mode_to_delta(
-                                mesh=self.mesh,
-                                global_params=self.global_params,
-                                positions=positions,
-                                fixed_mask_in=fixed_mask_in,
-                                delta_in=delta_in_trial,
-                            )
-                        )
-                        self._last_inner_coupled_update_mode_stats = dict(
-                            update_mode_stats
-                        )
-                        trial_in, trial_out = build_leaflet_trial_tilts(
-                            base_in=tilts_in,
-                            base_out=tilts_out,
-                            delta_in=delta_in_trial,
-                            delta_out=step * dir_out,
-                            normals=normals,
-                            fixed_mask_in=fixed_mask_in,
-                            fixed_mask_out=fixed_mask_out,
-                            fixed_vals_in=tilt_fixed_vals_in,
-                            fixed_vals_out=tilt_fixed_vals_out,
-                        )
-                        E1 = self._compute_tilt_dependent_energy_with_leaflet_tilts(
-                            positions=positions,
-                            tilts_in=trial_in,
-                            tilts_out=trial_out,
-                            grad_dummy=grad_dummy,
-                            tilt_vertex_areas_in=tilt_vertex_areas_in,
-                            tilt_vertex_areas_out=tilt_vertex_areas_out,
-                        )
-                        if E1 <= E0:
-                            tilts_in = trial_in
-                            tilts_out = trial_out
-                            E0 = E1
-                            accepted = True
-                            break
-                        step *= 0.5
-                        if step < 1e-16:
-                            break
-
-                    if not accepted:
-                        break
-
-                    accepted_steps += 1
-                    if (
-                        projection_cadence == "per_step"
-                        and (accepted_steps % projection_interval) == 0
-                    ):
-                        _refresh_tilts_from_constraints_and_project()
-
-                    E0, gnorm = _leaflet_tilt_gradients()
-                    if gnorm == 0.0 or (tol > 0.0 and gnorm < tol):
-                        break
-
-                    res_in = -tilt_in_grad
-                    res_out = -tilt_out_grad
-                    if M_inv_in is not None:
-                        z_in = res_in * M_inv_in[:, None]
-                    else:
-                        z_in = res_in
-                    if M_inv_out is not None:
-                        z_out = res_out * M_inv_out[:, None]
-                    else:
-                        z_out = res_out
-                    rz_new = float(np.sum(res_in * z_in) + np.sum(res_out * z_out))
-                    if rz_old == 0.0:
-                        break
-                    beta = rz_new / rz_old
-                    dir_in = z_in + beta * dir_in
-                    dir_out = z_out + beta * dir_out
-                    rz_old = rz_new
-
-            if accepted_steps > 0:
-                if projection_cadence == "per_pass":
-                    _refresh_tilts_from_constraints_and_project()
-                elif (accepted_steps % projection_interval) != 0:
-                    _refresh_tilts_from_constraints_and_project()
-
-            loss_outer_far_ratio = (
-                float(
-                    projection_norm_loss_outer_far
-                    / max(projection_norm_ref_outer_far, 1.0e-18)
-                )
-                if projection_norm_ref_outer_far > 0.0
-                else 0.0
-            )
-            self._last_tilt_projection_stats = {
-                "projection_cadence": str(projection_cadence),
-                "projection_interval": int(projection_interval),
-                "projection_apply_count": int(projection_apply_count),
-                "tilt_projection_norm_loss_outer_far": float(loss_outer_far_ratio),
-            }
-
-        self._set_leaflet_tilts_from_arrays_fast(tilts_in, tilts_out)
-        _apply_scaffold_interface_shape_projection()
-        if hasattr(self.constraint_manager, "enforce_tilt_constraints"):
-            self.constraint_manager.enforce_tilt_constraints(
-                self.mesh, global_params=self.global_params
-            )
 
     def refresh_modules(self):
         """Re-load energy and constraint modules from the current mesh state."""
@@ -2009,6 +1218,8 @@ class Minimizer:
         self._has_enforceable_constraints = any(
             hasattr(mod, "enforce_constraint") for mod in self.constraint_modules
         )
+        self._tilt_relaxation_manager.mesh = self.mesh
+        self._tilt_relaxation_manager.constraint_modules = self.constraint_modules
         logger.info(
             f"Minimizer modules refreshed: {len(self.energy_modules)} energy, {len(self.constraint_modules)} constraint."
         )
